Deduplicate fetchWithTimeout helper
